debug

+++ client/src/main/scala/com/neo/sk/carnie/controller/BotController.scala
@@ -190,7 +190,7 @@ class BotController(player: PlayerInfoInClient,
     }
   }
 
-  def getAllImage : List[Array[Int]] = {
+  def getAllImage : List[Array[Byte]] = {
     layeredGameScene.layered.getAllImageData
   }
